pyproject.toml
[tool.poetry]
name = "twitter-sentiment-analysis"
version = "0.1.0"
description = "A hybrid system for monitoring Twitter sentiment about cryptocurrencies"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
openai = "^1.3.0"
weaviate-client = "^3.24.1"
python-telegram-bot = "^20.6"
pandas = "^2.1.0"
apscheduler = "^3.10.1"
python-dotenv = "^1.0.0"
pydantic = "^2.4.2"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
twitter-sentiment = "twitter_sentiment.main:cli"